Full Tailwind integration (#34)
* Remove additional css postprocessors
Sass and stylus should not be required when using Tailwind

* Replace main style file
Design system should be configured in tailwind config.

* Remove sass references

* Add base colors and fonts

* Simplify btn for now

* Convert scss to css

* Remove redundant skin class

* Add postcss nesting

* Refactor app store
I think it helps to be more explicit here. Also, we should use mutations if all we are doing is mutating the state. Actions should only be used to do something prior to mutating state, e.g. fetching data from an external source.